% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_clean.R
\name{meteo_clean}
\alias{meteo_clean}
\title{Limpiar y estandarizar columnas}
\usage{
meteo_clean(df)
}
\arguments{
\item{df}{Tibble devuelto por `siga_read()`.}
}
\value{
Tibble con nombres estandarizados y columna `fecha` como Date (si existe).
}
\description{
Limpiar y estandarizar columnas
}
