% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga_download.R
\name{siga_download}
\alias{siga_download}
\title{Descarga datos meteorol\enc{贸}{o}gicos de una URL externa}
\usage{
siga_download(url)
}
\arguments{
\item{url}{La URL completa del archivo CSV a descargar.}
}
\value{
Un objeto `data.frame` (tibble) con los datos descargados.
}
\description{
Esta funci\enc{贸}{o}n realiza una petici\enc{贸}{o}n HTTP a una URL y lee el contenido
como un archivo CSV directamente en R, manejando errores de conexi\enc{贸}{o}n.
}
\examples{
\dontrun{
# Ejemplo simulado:
# datos <- siga_download("http://siga.inta.gob.ar/datos/estacion123.csv")
# head(datos)
}
}
