% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga_read.R
\name{siga_read}
\alias{siga_read}
\title{Leer archivo meteorológico del SIGA}
\usage{
siga_read(path)
}
\arguments{
\item{path}{Ruta al archivo CSV.}
}
\value{
Tibble con los datos meteorológicos crudos.
}
\description{
Leer archivo meteorológico del SIGA
}
