% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga_read.R
\name{siga_read}
\alias{siga_read}
\title{Lee un archivo de datos meteorol\enc{ó}{o}gicos del sistema SIGA o similar}
\usage{
siga_read(path)
}
\arguments{
\item{path}{Ruta al archivo CSV.}
}
\value{
Tibble con los datos meteorol\enc{ó}{o}gicos crudos.
}
\description{
Lee un archivo de datos meteorol\enc{ó}{o}gicos del sistema SIGA o similar
}
\examples{
\dontrun{
# Necesita un archivo real
ruta_prueba <- system.file("extdata", "siga_test.csv", package = "paqueteMeteo")
if (file.exists(ruta_prueba)) {
  datos <- siga_read(ruta_prueba)
  print(head(datos))
}
}
}
