% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siga_read.R
\name{siga_read}
\alias{siga_read}
\title{Lee un archivo de datos meteorol\enc{ó}{o}gicos del sistema SIGA o similar}
\usage{
siga_read(path)
}
\arguments{
\item{path}{Ruta al archivo CSV.}
}
\value{
Tibble con los datos meteorol\enc{ó}{o}gicos crudos.
}
\description{
Lee un archivo de datos meteorol\enc{ó}{o}gicos del sistema SIGA o similar
}
